+++ packages/cli/src/index.ts
@@ -3,3 +3,4 @@
  */
 
 export { trainCommand } from './commands/train';
+export { checkCommand } from './commands/check';